Fix stored ssrf
